+++ js/main.js
@@ -666,7 +666,10 @@ const MM = (function () {
 			}
 
 			// Further implementation is done in the private method.
-			updateDom(module, updateOptions);
+			updateDom(module, updateOptions).then(function () {
+				// Once the update is complete and rendered, send a notification to the module that the DOM has been updated
+				sendNotification("DOM_OBJECTS_UPDATED", null, null, module);
+			});
 		},
 
 		/**